:memo: Improve / fix some Spanish translations
I understand that some of these may come from a Spanish dialect – I'm open to discuss solutions, since I understand that from people of X dialect it might not be acceptable to be treated using Y dialect.

Also, I made consistent the usage of su / tu, usted / ti. I personally find it excessive / old / boring to be treated with 'usted' in a modern context, but you might not hold the same opinion. Again, open to discuss :).