Separate redux and rendering implementations
Wraps the process of creating a Redux store and decorates it with the
appropriate history module and redux enhancer for the appropriate
environment,

Creates two different rendering functions for react-dom and
react-dom/server. These functions expect to be passed a store,
and they will handle the rest based on the history.