Update asset dependency technique
This replaces everywhere in the frontend codebase where references
are made to image asset paths. It uses a webpack technique called
require contexts to build a map of asset urls by their source code
filename.  Webpack will generate hash based filenames for these assets
as part of the build, so this ensures that references stay in tact
without the developer needing to think about it.